I realized that I was setting all of the JSON values as strings which is likely suboptimal. I've moved away from the "GenerateMessage" method and created classes for each of the JSON objects that the server needs to instantiate and push to clients. This will also allow for more flexibility in the future if needed.

+++ SwinecideServer.cs
@@ -166,10 +166,10 @@ namespace SwinecideServer
                                 else if (msgDict["msgType"] == "RequestEntity") {
                                     // Find related match, relay message.
                                     idToPlayerDictionary[wsToIdDictionary[ws]].currentMatch.SendToOpponent(ws, msg);
-                                    long type = msgDict["type"];
-                                    long x = msgDict["location"]["x"];
-                                    long y = msgDict["location"]["y"];
-                                    long parentId = msgDict["parent_id"];
+                                    int type = (int) msgDict["type"];
+                                    int x = (int) msgDict["location"]["x"];
+                                    int y = (int) msgDict["location"]["y"];
+                                    int parentId = (int) msgDict["parent_id"];
                                     idToPlayerDictionary[wsToIdDictionary[ws]].currentMatch.EntityRequested(ws, type, x, y, parentId);
                                 }
                                 else
